Adds reboot into rescue mode
There were no usage instructions on how to reboot an instance into
the rescue mode. So, it was decided to add the description of
$ nova rescue SERVER and $ nova unrescue SERVER
commands to the user guide (Reboot an instance section).

As it is much related to the bug
https://bugs.launchpad.net/openstack-manuals/+bug/1346793
it also adds a note, informing the user about the restricted
number of available actions for the instance in running
in the rescue mode.

Besides, it adds --rescue_image_ref

Change-Id: I8818d4d7430c26619cd7bfc364c0f3d19f174152
Partial-Bug: 1346793
Closes-Bug: 1306174