Clarify Swift's implementation definition in the Install Guide page
The current definition of Swift's implementation in the Install
Guide page is confusing and, therefore, incomplete. The root of
the confusion is as it follows: "Its implementation is not like
a file server with mountable directories.", which doesn't
actually explain what it is like. Clarify this concept adding
information about Swift's implementation.

Patch Set 2: Rewording of paragraph.
Patch Set 3: Small adjustment regarding grammar issues.

Change-Id: I9bca8bd1df758e5de1201231efb3c245f1cfe9a4
Closes-Bug: #1412456