Support for ext4 as default for ephemeral disks
Currently, ext4 is supported as default for all Linux.
Also, for non-linux and non-windows OSs the default is vfat
(used to be ext3).

There is still support for ext3 as the default if some
deployers do choose to keep that as the default, by using either
virt_mkfs or default_ephemeral_format configuration option.

Change-Id: I1b315043d0738bc4eec6739ad82be52e395f8d58
Closes-Bug: #1415407