Testing including content based on meta definitions
- To try it yourself, use:
sphinx-build -t admin_only \
doc/playground-user-guide/source/ \
doc/playground-user-guide/build/html/.
- Adds a simple scope.py extension, is this the right place for the extension
to be maintained, or should it go to openstack-doc-tools repo?
- This approach is a file-by-file approach for conditionals and will not do
conditional output based on a smaller chunk, such as a note or paragraph. Is
this a good approach for the RST migration? I have also investigated smaller
chunks with the .. only:: directive but it is wonky with non-para content such
as admonitions, tables, list items.

Change-Id: Ia9a44ea45319c97073ab4de55320c1851def232e