Adds the Create a backup section (nova, CLI)
Images now have an `image_type` associated with them, which can be
snapshot or the backup type, like 'daily' or 'weekly'.

If the image_type is backup-like, then the rotation factor can be
included, and that causes the oldest backups that exceed the
rotation factor to be deleted.

This patch explains the usage of the nova backup command,
and the rotation mechanisms.

As Dashboard does not provide a user with this possibility,
so, the section is included to the Manage images section
which is related to CLI only.

Change-Id: I58ad41375ae5724b90d3c993395f318f8417e498
Closes-Bug: #1069986