Add migrated content for trove CLI chapter
This commit covers the overview section of
the original file section_cli_trove.xml.

This is a relatively large XML file, so I am
breaking it into separate RST files for each
sub-section.

The RST files for the sub-sections will be
present in future commits as I convert them.

Change-Id: I55f551feb6285b1991917f76dfd0ef8780ba13d5
Partial-Bug: #1419134