Reference to volume scheduling removed from Compute section
As of Juno, the config guide still claims that the Nova scheduler takes 
care of volume scheduling requests. This functionality has been in Cinder 
for years. Similarly, the MultiScheduler mentioned in the config guide 
dates from the times when Nova managed volumes and is now obsolete.

Fix: In the sentence "Compute uses the nova-scheduler service to 
determine how to dispatch compute and volume requests", removed the words 
"and volume". Also removed the line 
"scheduler_driver=nova.scheduler.multi.MultiScheduler" from the nova.conf 
example.

Closes-Bug: #1423788
Change-Id: I2046c81a730083891a7d1107aa754232fcfa9e61