Add links in HOT Guide to collections of existing templates.
There is not yet a central repository for Heat templates, but in
the mean time a list of links to collections of existing Heat
templates will be very helpful. This change adds one page to the
HOT Guide with links to several repositories. The latest patchset
should clear up the tox checks.

Closes-Bug: #1425581

Change-Id: I1949a8379d1d598cf399d37b4ddb87b549cfa7e6