Unify checkbuild and publishdocs
The two tox environments were large and mostly duplicated, use a shell
script instead and use the build-rst.sh script to simplify building and
publishing.

For copying to /kilo, use the tools/copy-to-branch.sh script instead of
the simple commands since that is more robust. Merge that script into
this script, no need for a second script.

Update requirements: We need 0.27 at least to not have to delete the
Debian Install Guide, sync also pbr with global-requirements.txt.

Change-Id: Ieba77b4e3d0bbe58db041206d9725c523b3bb43a