Fix kilo publishing
Fix a problem in publishdocs.sh: The copying of draft to kilo
copied the files to publish-docs/kilo/draft instead of to
publish-docs/kilo/ since the directory existed already.

Make this robust, always create the directory and copy the content of
it. Also remove draft-index.html, we don't need it in the kilo
directory.

Change-Id: I63e633dd4ecd4fd6ada74ca312d8cb6cb6a8fc28