Add launchpad tag used by "Report a bug"
"Report a bug" automatically populates bugs on launchpad with
information about the manual. This additional config setting
will allow "Report a bug" to automatically tag the bug with the
respective manual.

Add this for admin-cloud-rst like it was done for the other guides.

Also, decrease version number to 0.9, it's not a 1.0 release yet.

Change-Id: I4555cc28ef30527e4349da5c4928ba265a19bb23