Dashboard system requirements for the install guide
Because horizon is deployed on the controller node from Juno and Kilo onwards, this
patch removes the system requirements section, which has documentation
for deploying horizon on a server other than the controller node.

It moves any important information from the system requirements chapter to the
introduction and concluding paragraphs of the Horizon chapter.

Change-Id: I653c12c3e895111d08ecf8fef017e82fca2f909d
backport: juno kilo
Closes-Bug: #1426451