[install-guide-rst] Set up conditionals for distros
Add support for conditionals for the Install Guide.

This uses the tags:
* rdo
* obs
* ubuntu

Note that this is changed from the spec since sphinx does not accept a
"-". Also, I changed from opensuse to suse for brevity.

Adds new tox target so that "tox -e install-guides" builds only the
RST Install Guide.

implements bp installguide-liberty

Change-Id: Ibdfbc7ff294e56f52cd6f4ef59ceeb7b6af3e28a