Add Dell Storage Center Cinder extra spec info
The volume type extra spec storagetype:storageprofile
can now be used to allow volumes to be created with
Storage Profiles other than the default.

This patch adds documentation around what this feature
is and how to configure it using the extra specs.

This documents a new feature for Liberty.

Backport: none
Change-Id: I042a8db4b53ae7b5ab0f136ceeb6e8b09734c5e1
Depends-On: Icf76fceca5a0ae20bb08b276b0c41ef6cdb31087