DiskImageBuilder document link is elements, change word.
Chapter 6. Tool support for image creation in OpenStack Virtual Machine
Image Guide  - current - DiskImageBuilder link is wrong

In sentence "This example creates the image as before, but for arm architecture.
More elements are available in the git source directory and documented in the
diskimage-builder documentation. " the hyperlink referred by "diskimage-builder "
is wrong.

It's http://docs.openstack.org/developer/diskimage-builder/elements.html
(pointing to a section of the documentation) but should be
http://docs.openstack.org/developer/diskimage-builder/
(pointing to the documentation main page)

change word "diskimage-builder documentation" to "diskimage-builder elements
documentation."

Change-Id: I7deae6416b640ae96fddbcc0fbf469f72d5fd285
Closes-Bug: #1473040