Fixed text within callout modals
In the Installation Guide for Ubuntu 14.04  - kilo, under section Architecture > Overview,
glossary call-outs in table 1.1 for "Dashboard" and "Identity Service" are missing descriptions.
The <glossterm> string for "Dashboard" and "Identity service"  did not match the exact term used
in the glossary-terms.xml. The fix matched the correct term in the ch_getstart.xml to
glossary-terms.xml.

Closes-Bug: #1478390
Change-Id: Iefccc4f03f4c0ec9c61cc7f261a45b1ad7effe5c