Updates chown example of keystone conf file to entire directory
Changes the chown example to set the ownership of /etc/keystone
to both the user and group of keystone. Without this I was seeing
"ERROR: openstack Internal Server Error (HTTP 500)".

Note, when I went through the install I did not need to run an
additional "keystone-manage db_sync" as mentioned in the bug.
That step is documented earlier in the document.

Closes-Bug: #1482134
Change-Id: Ib2a7c07f077fba3a1c43a6f4fbfa671d321f6909