Correct Link to Launch Instances in Dashboard in Cloud Administrator Guide
In the Admin Guide for Cloud, under the Dashboard subcategory, the link
used as a reference to launch instances with the dashboard does not point
to the correct section in the End-User Guide.

This fix will update the current link to point to the correct section in
the End-User Guide.

Change-Id: I0c8db1f4b27286a739d32c444f3b94681b42f637
Closes-Bug: #1489133