Add Deployment section with Kolla
There are several deployment tools in OpenStack including Kolla, Tripleo,
OpenStack Ansible, OpenStack Chef, and OpenStack Puppet.  Add a section
for deployment tooling so the relevant projects can add documentation
for the respective projects.

Since the author of this patch is mainly interested in highlighting the
documntation of Kolla, start out with a Kolla entry in the Deployment
section.

Change-Id: Ie792bc5d04246edcdc4ab997e5da38f3813ddde6