Rework install guide build tool
We have two index files for Debian and other distributions.
On the other hand, when the tool generates translation resources,
uses only one index file, Debian index as of now.
Therefore, translation for index files is applied
to only one index file.
For example, Translation for Ubuntu index file
are not applied for Ubuntu Installation Guide.

* Unify index files into one file
* Process toctree for each distribution dynamically

Change-Id: I3abcd7da3569bb8e9edbc02574dcb74c12752ede
Closes-Bug: #1507861