Fix incorrect ordering of install packaging section
Change https://review.openstack.org/#/c/237899/ reordered various
aspects of the environment sections in the install guide and
inadvertently re-ordered the packaging section.

As reported by Timo Henriksson, after installing the repositories,
the package list must be updated before clients can be installed.
If this is not done, incorrect and incompatible versions will be used.

Change-Id: Ifac6d788105f535cdddff3161d0691d50cc12d62
Closes-Bug: 1508409