[install] link to the Liberty config reference
At the moment in "Add the Image service" we link to the trunk version of
the config reference which redirects to the Liberty config reference at
the moment. We should directly link the Libery config reference to not
use the Mitaka config reference with the Liberty install guide in the
future.

Change-Id: I4fbcf8791ecde078a8ce9dd7a46947c6dd7ac5c0