Add general reference for IBM Style Guide
The IBM Style Guide is used not only for spelling but in general, add it
as general guideline.

Original wiki content was "This page lists conventions that the Docs
team follows for writing documentation. Additionally, the IBM Style
Guide is used as reference. In case of conflicts between these two, the
wiki convention pages takes precedence."

Change-Id: Ic308841a3ddb9551c2b80ea237ec587165d3c59d