Update confusing note
The note saying that services other than Identity assign meaning to
roles is confusing. Maybe the intent was to say that the other services
have a policy file that defines what operations the roles can do.
Before the v3 API, keystone didn't support RBAC, but now with v3 it
supports RBAC like all the other services.

The note is updated so that hopefully it will now provide usable
information.

Change-Id: I9e57c4244f2e57aedc2f0dffbdf38ee07eb67473