[networking] Use explicit reference for section link
In sphinx, we can use a reference by a title name, but it does
not work when a title is translated. Instead we should use
an explicit reference for section link.
It is required to avoid an error of translation build.

Change-Id: I0ec8109088946ec5bc0b63d2e09ec021f3b5307c
(cherry picked from commit ff839ae2d05339675a68abf61971ade79a7e838f)