[user-guide] Fix markup for easier translation
In the previous version, the string contains a backslash after
a closing ** markup. It is rendered in English version, but
it cannot be rendered properly even if we try to use a similar
markup in translations. (The only way I found to make Japanese
translation rendered properly is to use the proposed way.)

The proposed string changes nothing in English version
and make translators easy to handle the string.
There seems no meaning in using a backslash here.

Change-Id: If1ad1aedb235e216c66ab0a4f8307b873c69f1e4