Remove some useless configuration command
A symbolic link which is /etc/neutron/plugin.ini pointing to
/etc/neutron/plugins/ml2/ml2_conf.ini is told to create when
we install and configure the networking service in the
compute node.
But it's no need to be created, because there is no
'/etc/neutron/plugins/ml2/ml2_conf.ini' file and
the ML2 plug-in is installed in the controller node.
Thus we should remove this command and skip this step.
Backport: Liberty

Change-Id: I76feb15477ffd2d65cc8eb46c7472b6de830105f
Close-bug: #1525075
(cherry picked from commit 76e5f6859b1e99fffe39f917f253da0a5ac67912)