Update translation setup
Since the only guide that uses the common directory is now the
cli-guide and we do not translate that one, disable translation of the
common directory as well.

Remove now obsolete setting of DIRECTORIES for zh_CN since those
were only relevant for the DocBook version of these.

Adopt comment for contributor-guide to move it under the do not
translate section, translation this content is not a focus right now.

Change-Id: I0ea32b15c7debba383c0840d1051a8812df2cdee