Add manila install guide documentation
It is an effort to document manila install guide.
Only Ubuntu distro has been properly updated until now.
Liberty packages have been used for the tests, but it can be
changed to mitaka as soon as they are available. Thanks in
advance for any feedback about this initial effort.

Change-Id: I686bdda2ee27de4e5c43bb8d95c0126c8c300558
Implements: blueprint create-manila-install-guide