Add Network IP Availability API Extention
There was api merged for Network IP Availaibility [1]. This change
basically implements new extension that can be used to check network
usage stats of all the networks or specific network created by neutron.
There is also patch [2] upstream and close to merge for client commands

This patch will add how to configure it and commands used to get that
information.

[1]. https://review.openstack.org/#/c/212955/
[2]. https://review.openstack.org/#/c/269926/

Closes-Bug: #1551907
Change-Id: I5a6df488147240c11a4bad7f03320469bc8a3be6