Add missing setting for cache
The install guide instructs how to use memcached for a session
store for Horizon. It configures the caches, but does not actually
tell Horizon to use it for sessions. The default is signed-cookies
for sessions not caching.

This patch adds the setting to tell Horizon to use the cache for
sessions.

Change-Id: I7cd2add9df5dd699879dbdd1406495dc03f5267b