Remove parallel building
Parallel building is not working reliable with errors, I just had a
problem were the script run successfully but should have failed.
Let's be on the safe side...

Also fix  "WARNING: term not in glossary: fqdn" and
"WARNING: unknown document: example-architectures".

This shows that the parallel building hit problems, otherwise this
would not show up now.

Change-Id: I5de3446aceb0ba6aa658b09987f7ee9d4b116229