fix ring building instructions
Previous a different zone was being assigned to each device,
which is not how a typical deployment works. Zones are usually
assigned to correlate to a failure domain at a higher level.
For example a node or a rack.

In these instructions I changed to use just two zones, one for
each node (ip address), I think this gives a better
representation of how zones should be assigned.

backport: liberty

Change-Id: I17bb3d59089af59fb140ed096d3d0b42ad149ade
Signed-off-by: Thiago da Silva <thiago@redhat.com>