New heading style (plus)
A recent change in openstack-doc-tools caused the heading style
to change. This changeset updates the clients where that
change is present and where small additional other changes exist
(such as minor changes in formatting options, etc.) namely:

    barbican
    designate
    glance
    heat
    mistral
    senlin

This is done to ensure consistency as recent client updates
have already switched to this mode.

Change-Id: I45d40eb7a07b38ce3c0241094c27195f253a4893