[networking] remove usage of external reference for translation
Translation support

Sphinx allows us to use external references as follows, but labels
cannot be translated. If we translate labels, we will get dangled
links in the generated translated document.
This patch replaces such external references in the networking guide
with the style of `User guide <http://docs.openstack.org/user-guide/>`_.

----
This paragraph contains the link to `User guide`_.

 .. _`User guide`: http://docs.openstack.org/user-guide/
----

Change-Id: I80b590bd980300420f55128669219e26ad7ec2de