[admin-guide] Fix IPs in Share Replication documentation
The IP Addresses used in the documentation of the replication
feature recently introduced in Shared File Systems service could
be perceived to be mis-configuration.

Replace them with randomly generated private IPs and clarify
the usage of these IPs.

Also separate the 'access rules' and 'snapshots' sections
for better readability.

TrivialFix

Change-Id: I00b7df8472950a18f28dda5bacf23265b17047d7