[install] do not install/start/enable nova-cert service
According to http://docs.openstack.org/developer/nova/man/nova-cert.html:

---snip---
nova-cert is a server daemon that serves the Nova Cert service for X509
certificates. Used to generate certificates for euca-bundle-image. Only
needed for EC2 API.
---snap---

Because we do not install the EC2 API we do not need the nova-cert
service.

Change-Id: I0c27c89fb6f6b982ce227273e040216828b40135