Add note about real unit of burst value in Neutron API
There is mistake in Neutron's API in name of burst parameter. Name of
this parameter may suggest that burst is given in kilobits per second
but in fact it is given in kilobits.
Due to problems with backward compatibility with fixing that for now
we are adding only note to docs about that. Bug should be fixed when
Neutron will support microversioning.

Change-Id: I303ec63f98cd76f566d37d63f0db52cd51905e24
Related-Bug: #1507761