Clarify use of cinder manage-snapshot command
The snapshot in the 'cinder mansage-snapshot' command refers
to a snapshot of a volume, but the documentation mentions
"A snapshot is an image created from a running instance". This is
misleading since that indicates that it is a Nova snapshot.
Also, it should be explicit that the volume refered to in the command
is the parent of the snapshot.

Change-Id: Ia958b571cbb59d5046661ff85f8545f4f56a140b