[user-guide] Reorganize SDK section
The current structure of SDK section is not well listed in the
toc tree. This commit moves the overview in sdk.rst to a sub page
sdk_overview.rst so that all subsections in SDK section are displayed
in the toc.

Also moves 'Create a legacy client object' page to upper level.
Previously this page is listed as a bullet in the authenticate page
so it was small. This change makes the page more highlighed.

Change-Id: I6414b748a77a6375423da38e52113b36359e5f7d