Correct note about 3PAR iSCSI address config
The HPE 3PAR Fibre Channel and iSCSI driver configuration reference
currently indicates that address ports can be specified in the
hpe3par_iscsi_ips option. Change this note to indicate that this
option should simply be a comma separated list of IPs. The 3PAR array
does not allow the default port 3260 to be changed.

Change-Id: I36523a042753fc899923bc86efb51fc74884ca2c
Closes-Bug: #1601892