Remove debconf from nova-compute-node install
As per the Keystone install, this removes the parts which
are referencing Debconf setup, and remove lots of conditionals.

Note that intentionally, the part that activates use_neutron = true
and firewall_driver is also removed from the Debian isntall guide,
as this is already the default in the Debian package.

Change-Id: Ia0b2c4875052bf08045d28eb508f7e76a3a484d8