[networking] Default security group per project not network
The intro networking guide incorrectly described the default security
group as belonging to a network, rather than a project. If this where
the case, a project would not be able to modify the default security
group for an instance attached to a shared network or another project
could modify the security groups for the instances belonging to the
first project.

backport: mitaka
Change-Id: I6587728bb4c86451fc88d796537ebcae6ed72997