Add info about ports to sf_svip config entry
The SolidFire config ref was updated to include the requirement
of specifying sf_svip in the config file when using vlans,
however the update doesn't mention the fact that when specifying
this option the admin MUST also include the port number in the ip.

So rather than just "10.5.3.4" it must be of the form "10.5.3.4:3260".

The mitaka release (forward) will automatically append a default
port number in the driver, but that's not available in Liberty.

This patch just updates the Liberty docs to point out that port
number is required in the argument.

Change-Id: I25b4432df63831a42921ed28dab1699f321562a0
Closes-Bug: #1614243