[networking] Migrate FWaaS docs from admin to networking guide
With the introduction of FWaaS v2, it is time to revamp the
documentation for FireWall as a Service and bring it up to speed with
current standards.  As part of that, the documentation is being
relocated from the Admin Guide to the Networking Guide.  The
documentation is also being sliced into FWaaS v1 and FWaaS v2 scenarios,
so that they can be added in the correct places in the documentation
tree.

To do in future patch:
- The FWaaS v2 scenario needs to be updated to show exactly how FWaaS
  will be used with API calls.
- The 'feature comparison' section needs to be expanded to show all of
  the points of comparison between v1 and v2.

Change-Id: I8a1635743d8df22970405193180916bd438605bb