[install] Incorporate bootstrap into keystone content
Since the keystone-manage bootstrap command has been implemented we
can refactor our installation process to use that, instead of
creating the Identity service and endpoints manually with the
ADMIN_TOKEN.

This also includes moving a bunch of the information in
keystone-services.rst to the get-started-identity.rst guide, since
creating the keystone service and endpoints manually is no longer
required (bootstrap does this for us now). The remaining information
in keystone-services.rst for exporting environment variables to use
the ``openstack`` client were moved to the keystone-install.rst
document since that's essentially the last step of keystone's
installation and configuration.

Change-Id: I6be3f286daf968dddf808f0204f2b317a5f09a59