LB is only pingable after the listener is created
Under the heading "Building an LBaaS v2 load balancer", step 4. directs
the reader to "Verify that the load balancer is responding to pings
before moving further". However, the loadbalancer is not actually
pingable until step 1. of "Adding an HTTP listener" is completed.

This can lead a reader to think their setup is not working and cause
them not to proceed further when in fact they simply need to add a
listener.

Change-Id: I2a5a5d170ba4e2e22d0a4d1e3bc4b5cc37d0dfe7
Closes-Bug: #1643831