Show team and repo badges on README
This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

b'https://gist.github.com/326db469cfd3580aa8fe949a671348fd\n'

Change-Id: I28487d4403825604a6c7e298ddf06e2dde4ee5df