Encrypted volumes doc should use 256 bit key
The sample command uses a key size of 512 bits for an xts cipher.
While this is actually a valid key size and cipher for a typical
cryptsetup configuration, a key size of 512 bits does not play
nicely with the key managers, since generally they do not exceed
256 bits.

Changed the key from 512 to 256 becacuse standard key manager like
any of the production-ready key managers OpenStack uses, mostly do
not support 512 bit AES keys.

Change-Id: I795d28120f86fa22b2eaeee44c6dc5c4aa40c8c3
Closes-Bug: 1616239