prepare to use reno from Pike onwards
A stated goal from the Atlanta PTG this week was to use reno for our
release notes from Pike onwards.  This requires reno to support
configurable sections, which have been added by:

    https://review.openstack.org/#/c/436639

Once this is merged into reno, this commit configures openstack-manuals
so that reno will work with it, and a build of releasenotes/ will
automatically include a reno-generated report via reno's sphinx
extension.

Depends-on: I914572c6a07ca81c54965b4b5a6b6aba50b3a787
Change-Id: Id029f462fb553b3429629b90585e2790734f1bb9