add a test page to the build structure for release notes
The release notes build needs a page that uses an unconstrained
"release-notes" directive so that the build in the CI system includes
the patch being proposed. In cases where a branch name is specified,
only changes already merged into that branch (not submitted to it) are
included.

This patch adds the missing file using the convention used
elsewhere (calling it unreleased.rst). It leaves the release-notes
directive in index.rst because I'm not sure whether that's where you
want "published" notes to appear.

It removes the branch setting in config.yaml so that when reno is not
given a branch it uses the current branch by default. This allows
patches under test to be included in the build, so that formatting
errors in the notes will cause the build to break before the patches are
merged.

It also reorganizes the order of the release notes pages to match the
order used by convention and expected by the repository branching script
so that when this repo is branched the new release note page for the
series can be added automatically by the branch script.

Change-Id: Ibccdac675915ab3535f2b2fbc1f4daee6990048b
Signed-off-by: Doug Hellmann <doug@doughellmann.com>

squash

Change-Id: Ie795176a7e860e10acb2cfa2e9d21c1d08ba8a48
Signed-off-by: Doug Hellmann <doug@doughellmann.com>