fix update-cinder-doc release note
reno was ignoring the note because it was not in a known section, so
move it to the "config" section to reflect the fact that the changes
were in the configuration guide.

The structure of the note file was also incorrect, so that is fixed at
the same time to ensure that the YAML parser sees a list item with a
literal text block containing restructuredtext.

Change-Id: Id8977cb57adc6813140707da8ac501c544959d54
Signed-off-by: Doug Hellmann <doug@doughellmann.com>