Disable branch root-marker
At branch creation post jobs are run and they create the root-marker -
but we're not ready yet. Uncomment these for now.

With stable/ocata creation, these root-markers now do not allow us to
publish new draft guides. We need to manually remove root-marker files
from the servers to update them. Note: root-marker protect directories
so that they are not deleted.

Change-Id: I05ad0721404d476cc1a72a8a9cbb177d206cba27