Don't use 'we'
The document used "The admin or other privileged
user must include this option during network
creation or add it later. In this case, we can
add it to the existing provider provider network.".

To use "we" in this case should be avoided.

Change-Id: Idae1b9459bbbc6452891e9864362903653927457
Closes-Bug: #1679584
(cherry picked from commit 5c91e66384fa16f1d1ac869682208f84c0a11d63)