Fix nova placement API guide for OBS
The nova package in SUSE requires the user to explicitly enable the
Apache vhost for the nova placement API, it is not a part of the
nova-api systemd service nor does it default to enabled. This patch adds
the necessary instructions to ensure the placement API is running on a
SUSE host.

backport: ocata
Change-Id: I104285439c9d46ded4dac0a773648a314741a3e3
(cherry picked from commit 922a2adee2734e477e235468305ffad037041927)