Update nova-compute-install steps for multi-cell support
Similar to the nova change with the same change ID, we
need to update the install guide for Pike due to some
limitations in the compute API when supporting multiple
cells. The hypervisor list command will not work before the
hosts are discovered, and with the steps as they are now in the
guide, you wait to discover the hosts until the hypervisors
show up, which is a chicken-and-egg problem.

The solution proposed is to use the openstack compute service
list command instead, which does not have the same host
discovery pre-requisite limitation. The nova cells install guide
docs are also being updated in the nova tree for the same issue.

Kolla is also making a similar change for their Pike deployment
scripts:

Id061e8039e72de77a04c51657705457193da2d0f

Change-Id: If2baab40c2e2a3de20e561bba50688d615b002ef
Closes-Bug: #1682060