Update and expand cells config doc warning
With cells v2 in Nova we have had to start going through
various docs both within and outside Nova to clarify when
something is talking about cells v1 vs cells v2. The cells
compute config reference document is all about cells v1.

This change updates the wording in the warning and moves it
to the top of the page so it's clear from the start that
cells v1 is considered experimental and not recommended for
new deployments, and is being replaced with cells v2. A link
for more information about cells v2 is provided in the warning.

Change-Id: I20204efc509237b3c5bacb22abed815a2b1e49ec
Closes-Bug: #1683416