Add module apply usage guide
This guide can be useful for those who need to apply software
modules to their databases like updates, external packages,
or probably the most common case, licensing.

This guide would have not been possible without the help of
Peter, he wrote this guide for me when I asked for help in
a review. I've been sitting on it for some time and made
all the necessary formatting changes, minor edits, as well as
reperformed all the steps so the output is accurate.

Note: I was admin when making this guide so there are fields
that aren't normally visible to the user.

Change-Id: I784b344adb29382287052c1aedaf7f96656adb8b
Co-Authored-By: Peter Stachowski <peter@tesora.com>
Co-Authored-By: chenxing <chen.xing@99cloud.net>
Closes-Bug: #1669050