fix pdf build
The PDF build does not include content multiple times if the same file
is included in a toctree more than once. That means we need to
restructure the guide to handle the common parts differently. This
approach merges some of the previously split sections back together
using inline prose to indicate where minor variations apply for
different operating systems but retaining separate files for cases where
the differences are significant.

Change-Id: I5d9ff549b05ca4ce54486719d70858589b8fcfa3
Depends-On: Ia750cb049c0f53a234ea70ce1f2bbbb7a2aa9454
Signed-off-by: Doug Hellmann <doug@doughellmann.com>