move /developer/.* out of the subdirectory
Move the project list up to the top level of the site and update the
links to point to the new locations of the projects.

Move the language bindings list to the top level of the site and update
the links.

Establish a redirect rule for /developer/$project/ to point to
/$project/latest/.

Change-Id: I8f6c6d8bcfff7c805c2912b76b49a9d74d49c6d5
Depends-On: Ia750cb049c0f53a234ea70ce1f2bbbb7a2aa9454
Signed-off-by: Doug Hellmann <doug@doughellmann.com>