Document AD nested groups support in keystone
Active Directory has a very specific mecahnism to handle nested groups. The
Identity LDAP supports nested groups in Acitve Directory as bug #1638603 has
been resolved. It has a fix released in both newton and ocata, but it is only
doucmented in Ocata. This change lists it in the description of LDAP
configuration options in the Newton docs.

Change-Id: Ieff311726dceaa63713fb85a3ba5fe57447aa494
Closes-Bug: #1700053