Distinguish between cloud-wide and per-service apis
The two are different use cases, and we have things that focus on the
cloud-wide story. Lead with those, then show docs for individual
services that users can fallback to if they need to.

Update docs and update this for all branches that your the project-data
for language-bindings (pike and queens only).

Change-Id: Ib1f7b7c77ec2b9536c4a03425694a6300025a6c3