[arch-guide] Updating EOL notice to index page
As of the Pike release, the documentation team determined they are
unable to continue maintaining the arch guide with the amount of
tech debt it produces. The intention is to continue to publish the
guide, and accept outside contributions, however the openstack-manuals
team is no longer solely responsible for the update(s) to the guide.

This notice will eventually be replaced by a banner that Jimmy McArthur
is currently working on.

Change-Id: I0bf146c12b36d4b652bb2f583dc7eba9335be69c