Add etcd3 to installation tutorial
Etcd is a base service [1] so it should be include in the installation
tutorial so that other services can assume that it is installed.

Right now, there is no package available for etcd3. This patch suggest
to download the released tarball in github [2] instead. The installation
steps are similar to the automated installation script in devstack [3].

[1] https://governance.openstack.org/tc/reference/base-services.html
[2] https://github.com/coreos/etcd/releases/tag/v3.1.7
[3] https://github.com/openstack-dev/devstack/blob/master/lib/etcd3

Change-Id: Ic532e5a01571900d60d51c3e94ba170e22a30e93