use macros to standardize redirect rules and tests
Define some macros for the standard cases like redirecting a
directory, redirecting everything under a path to a single page, and
redirecting everything under a path to a different prefix.

The new macros for redirecting a directory or a path also handle
trailing slashes in input paths more consistently than some of the old
rules.

Change-Id: I65579a70c03c62475878b2e1540c1927db81880e
Signed-off-by: Doug Hellmann <doug@doughellmann.com>