Raise on init, parse, render, or write fails
If jinja template initialization, parsing, rendering, or writing fails,
we used to sometimes log an error message, then sometimes continue
(ignoring the failure), sometimes reraise the exception, and sometimes
fail without raising the exception.  With this change, we always raise
the exception so we print the traceback and fail the build right away.

Change-Id: I04a36c4c8c69dbd08e9ea677983dbe00fa3a67a3