Update Deploy guides publication
OpenStack-Ansible and Kolla already have published deploy
guides, but it seems like they are not in the index page.

Updating project data for those projects should be enough to
get them generated into the
``https://docs.openstack.org/pike/deploy/`` index.

On top of that, OpenStack-Ansible also have a Queens
documentation published under "latest", so I am also adding it
to the list of the deploy guides. The names of those
deploy tools is also edited to match the names of Pike
deploy guides.

Closes-Bug: #1727779
Change-Id: I59b73eb3f5afc7ef5cdb64252d558b57ea26abf2