Fix deep_links macro in .htaccess
Because the trailing / is optional, the deep_links macro could match
multiple projects if one of their names is a prefix of the other(s)
(e.g. heat and heat-translator). It also means that completely bogus
paths will redirect to bogus locations - e.g. currently
https://docs.openstack.org/developer/nova-foo redirects to
https://docs.openstack.org/nova/latest/-foo

Avoid this by adding a negative-lookahead assertion, to ensure that the
path passed in by the caller of the macro is not followed by any non-/
character.

Change-Id: I9dcf5dccfa36a3b1895f4422deafcdb73b150cbd