link to the mitaka CLI guide
Replace the use of the project_guides.tmpl template in the mitaka
index file with inline versions that link to the mitaka-specific
pages.

Add a language-bindings.html page based on the one for ocata to link
to the user-guide and cli-reference for mitaka.

Add a projects.html page to link to the mitaka documentation for each
service project.

Change-Id: I173ecabeb3aa74a3d2ddc6b4ba898d881dcf752d
Signed-off-by: Doug Hellmann <doug@doughellmann.com>