ignore unofficial projects in latest.yaml
Update the template generator to ignore unofficial projects when
rendering for the current series. Ignoring, rather than throwing an
error, allows us to clean up the data without the build breaking.

To make the validation work for glance-store, a new field is needed to
set the 'deliverable-name', since it is different from the project
name used to build URLs.

Change-Id: I0876c93b49d57818d0b689fd03a0171a96e27547
Signed-off-by: Doug Hellmann <doug@doughellmann.com>