include governed projects in output
In 0ba83e5 we started trying to ignore unofficial projects and
introduced a bug so that all projects were dropped from the project
list as part of that evaluation. This patch fixes the bug by copying
the valid data from the raw values read in to the list of values being
passed to the template.

Change-Id: I4aeafe6d8116f3b032a5d339660878a76a10e2f3
Signed-off-by: Doug Hellmann <doug@doughellmann.com>