Fix the install guide buttons on small screens
This patch fixes the nesting on the rows on the install guide
index page, that were causing all of the links to be covered on
smaller screens.

Unfortunately the high line count change is really required to
make the change easy to understand, as otherwise the indentation makes
it very difficult to notice the incorrect nesting (and probably caused
the bug in the first place).

Closes-Bug: 1751810
Change-Id: Id4e7ddf9d46d6e1a7f81a88047bb6778453cc9b8