Add redirect for python-openstacksdk project rename
Add a macro to make generating the redirect and tests for a project
rename easy.

Redirect python-openstacksdk to the new name openstacksdk.

Because the .htaccess template generates redirects for all official
projects, this patch won't pass the tests until the governance change
with the rename is approved and merged, so that patch is included as a
depends-on.

Depends-On: https://review.openstack.org/555780/
Co-Authored-By: Monty Taylor <mordred@inaugust.com>
Change-Id: I97afa686cac31b2ef8beac1df49f98625dee703d