build the redirects for the project-deploy-guide dynamically
Pass the list of maintained series to the template environment so we
can use that list to generate all of the redirects to the
project deploy guide pages instead of having to update the list of
series in two places.

The newton series is handled as a special case because it exists but
is not considered maintained.

Change-Id: I10794819260f1edd54e3711258ab715910ccf20f
Signed-off-by: Doug Hellmann <doug@doughellmann.com>