add a flag to indicate which projects have guides
Some of the projects publish documentation but not using the project
guide structure. Instead of requiring all projects listed to have a
default guide, add a flag that defaults to True and that can be
changed in the project data file(s).

Also add a default field to the URLSettings structure.

Change-Id: Ie2b6abafde5d4be85941fd977762bcbdf891d440
Signed-off-by: Doug Hellmann <doug@doughellmann.com>